Replace unsigned int with NSUInteger.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileFoundation@6843 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/NSIndexPath+Etoile.h
@@ -8,15 +8,15 @@
 	Date:  September 2007
 	License:  Modified BSD (see COPYING)
  */
- 
+
 #import <Foundation/Foundation.h>
 
 
 @interface NSIndexPath (Etoile)
 + (NSIndexPath *) indexPath;
 + (NSIndexPath *) indexPathWithString: (NSString *)aPath;
-- (unsigned int) firstIndex;
-- (unsigned int) lastIndex;
+- (NSUInteger) firstIndex;
+- (NSUInteger) lastIndex;
 - (NSIndexPath *) indexPathByRemovingFirstIndex;
 - (NSString *) stringByJoiningIndexPathWithSeparator: (NSString *)separator;
 - (NSString *) stringValue;
